@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.17.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -28,53 +28,53 @@ environment:
 # versions available, run `flutter pub outdated`.
 dependencies:
   bottom_drawer: ^0.0.5
-  connectivity_plus: ^2.3.6
-  crypto: ^3.0.2
-  cupertino_icons: ^1.0.2
-  device_info_plus: ^4.1.2
+  connectivity_plus: ^6.0.3
+  crypto: ^3.0.3
+  cupertino_icons: ^1.0.8
+  device_info_plus: ^10.1.0
   diff_match_patch: ^0.4.1
-  dio: ^4.0.6
-  disable_battery_optimization: ^1.1.0+1
-  # easy_localization: ^3.0.1
-  easy_localization: 
-    git:
-      url: https://github.com/guillaumeboehm/easy_localization.git # For missing string fallback
-  enum_flag: ^1.0.1
+  dio: ^5.4.3+1
+  disable_battery_optimization: ^1.1.1
+  easy_localization: ^3.0.7
+  # easy_localization: 
+  #   git:
+  #     url: https://github.com/guillaumeboehm/easy_localization.git # For missing string fallback
+  enum_flag: ^1.0.2
   equatable: ^2.0.3
   flutter:
     sdk: flutter
-  # flutter_background_service: ^2.1.3
-  flutter_background_service:
-    path: /home/yoro/repos/flutter_background_service/packages/flutter_background_service # To avoid request on import
+  flutter_background_service: ^5.0.5
+  # flutter_background_service:
+  #   path: /home/yoro/repos/flutter_background_service/packages/flutter_background_service # To avoid request on import
   flutter_bloc: ^8.0.1
   flutter_dotenv: ^5.0.2
-  flutter_local_notifications: ^9.7.0
+  flutter_local_notifications: ^17.1.2
   flutter_localizations:
     sdk: flutter
-  flutter_svg: ^1.0.3
-  fuzzy: ^0.4.0-nullsafety.0
-  hive: ^2.1.0
+  flutter_svg: ^2.0.10+1
+  fuzzy: ^0.5.1
+  hive: ^2.2.3
   hive_flutter: ^1.1.0
-  launch_at_startup: ^0.1.7
-  objectbox: ^1.6.2
-  objectbox_flutter_libs: ^1.6.2
-  package_info_plus: ^1.4.2
-  permission_handler: ^10.0.0
-  # quick_notify: ^0.2.1
-  quick_notify: 
-    path: /home/yoro/repos/quick.flutter/packages/quick_notify # To avoid request on launch
-  sentry_flutter: ^6.9.0
+  launch_at_startup: ^0.3.0
+  # objectbox: ^4.0.0
+  # objectbox_flutter_libs: ^4.0.0
+  package_info_plus: ^8.0.0
+  permission_handler: ^11.3.1
+  quick_notify: ^0.2.1
+  # quick_notify: 
+  #   path: /home/yoro/repos/quick.flutter/packages/quick_notify # To avoid request on launch
+  sentry_flutter: ^8.2.0
   settings_ui: ^2.0.2
   shared_preferences: ^2.0.15
   system_tray: ^2.0.1
-  throttling: ^1.0.0
+  throttling: ^2.0.1
   universal_html: ^2.0.8
   universal_io: ^2.0.4
   universal_platform: ^1.0.0
   url_launcher: ^6.1.5
   url_strategy: ^0.2.0
   validators: ^3.0.0
-  window_manager: ^0.2.5
+  window_manager: ^0.3.9
   workmanager: ^0.5.0
 
 dev_dependencies:
@@ -82,9 +82,9 @@ dev_dependencies:
   build_runner: ^2.1.10
   flutter_test:
     sdk: flutter
-  hive_generator: ^1.1.2
-  objectbox_generator: ^1.6.2
-  very_good_analysis: ^3.0.0
+  hive_generator: ^2.0.1
+  # objectbox_generator: ^4.0.0
+  very_good_analysis: ^5.1.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
