@@ -1,2 +1,2 @@
-export 'package:flutter_memosync/src/services/storage/storage_hive.dart'
-    if (dart.library.io) 'package:flutter_memosync/src/services/storage/storage_objectbox.dart';
+export 'package:flutter_memosync/src/services/storage/storage_hive.dart';
+//     if (dart.library.io) 'package:flutter_memosync/src/services/storage/storage_objectbox.dart';
